requirements.txt
mmh3
bitarray
jellyfish
django
djangorestframework